requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
coverage==6.3.2
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.49.0
django-annoying==0.10.6
django-cleanup==6.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-filter==21.1
django-override-storage==0.3.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
oauthlib==3.2.0
Pillow==9.0.1
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
zipp==3.6.0